HIVE-28401: Drop redundant XML test report post-processing from CI pipeline (Stamatis Zampetakis reviewed by Alessandro Solimando)
The Maven Surefire Plugin (3.0.0-M4) does not store system-out and
system-err in the XML report for successful tests; the plugin stores
out/err in XML for skipped/failed tests.

The main intention of the xmlstarlet post-processing step is to remove
out/err for tests that run sucessfully since it can be very large and
most of the time we don't need to retain it.

The change leads to a slight change in behavior since now the out/err
from skipped tests will be retained. As a result the size of the
test-results.tgz archive will also increase (+1.5MB).

Nevertheless, the change is beneficial since it will avoid the
"Huge input lookup" problem, which is more detrimental for the CI
when it happens. In addition, it alleviates the burden to handle
out/err from Hive CI and leaves that responsibility to Surefire, that
is more appropriate.

Close apache/hive#5364